Change int to NSInteger to fix warnings.

+++ MGScopeBar.m
@@ -59,15 +59,15 @@
 @interface MGScopeBar (MGPrivateMethods)
 
 - (IBAction)scopeButtonClicked:(id)sender;
-- (NSButton *)getButtonForItem:(NSString *)identifier inGroup:(int)groupNumber; // returns relevant button/menu-item
-- (void)updateSelectedState:(BOOL)selected forItem:(NSString *)identifier inGroup:(int)groupNumber informDelegate:(BOOL)inform;
-- (NSButton *)buttonForItem:(NSString *)identifier inGroup:(int)groupNumber 
+- (NSButton *)getButtonForItem:(NSString *)identifier inGroup:(NSInteger)groupNumber; // returns relevant button/menu-item
+- (void)updateSelectedState:(BOOL)selected forItem:(NSString *)identifier inGroup:(NSInteger)groupNumber informDelegate:(BOOL)inform;
+- (NSButton *)buttonForItem:(NSString *)identifier inGroup:(NSInteger)groupNumber
 				  withTitle:(NSString *)title image:(NSImage *)image; // creates a new NSButton
-- (NSMenuItem *)menuItemForItem:(NSString *)identifier inGroup:(int)groupNumber 
+- (NSMenuItem *)menuItemForItem:(NSString *)identifier inGroup:(NSInteger)groupNumber
 					  withTitle:(NSString *)title image:(NSImage *)image; // creates a new NSMenuitem
 - (NSPopUpButton *)popupButtonForGroup:(NSDictionary *)group;
-- (void)setControl:(NSObject *)control forIdentifier:(NSString *)identifier inGroup:(int)groupNumber;
-- (void)updateMenuTitleForGroupAtIndex:(int)groupNumber;
+- (void)setControl:(NSObject *)control forIdentifier:(NSString *)identifier inGroup:(NSInteger)groupNumber;
+- (void)updateMenuTitleForGroupAtIndex:(NSInteger)groupNumber;
 
 @end
 
@@ -497,7 +497,7 @@
 			//NSLog(@"Got %@ - modifying groups %@", ((narrower) ? @"narrower" : @"wider"), NSStringFromRange(changedRange));
 			NSInteger nextXCoord = NSNotFound;
 			if (adjusting) {
-				for (int i = changedRange.location; i < NSMaxRange(changedRange); i++) {
+				for (NSUInteger i = changedRange.location; i < NSMaxRange(changedRange); i++) {
 					NSMutableDictionary *groupInfo = [_groups objectAtIndex:i];
 					
 					if (nextXCoord == NSNotFound) {
@@ -670,7 +670,7 @@
 }
 
 
-- (NSButton *)getButtonForItem:(NSString *)identifier inGroup:(int)groupNumber
+- (NSButton *)getButtonForItem:(NSString *)identifier inGroup:(NSInteger)groupNumber
 {
 	NSButton *button = nil;
 	NSArray *group = [_identifiers objectForKey:identifier];
@@ -685,7 +685,7 @@
 }
 
 
-- (NSButton *)buttonForItem:(NSString *)identifier inGroup:(int)groupNumber 
+- (NSButton *)buttonForItem:(NSString *)identifier inGroup:(NSInteger)groupNumber
 				  withTitle:(NSString *)title image:(NSImage *)image
 {
 	NSRect ctrlRect = NSMakeRect(0, 0, 50, 20); // arbitrary size; will be resized later.
@@ -811,7 +811,7 @@
 }
 
 
-- (void)updateMenuTitleForGroupAtIndex:(int)groupNumber
+- (void)updateMenuTitleForGroupAtIndex:(NSInteger)groupNumber
 {
 	// Ensure that this group's popup (if present) has the correct title,
 	// accounting for the group's selection-mode and selected item(s).
@@ -974,7 +974,7 @@
 }
 
 
-- (void)updateSelectedState:(BOOL)selected forItem:(NSString *)identifier inGroup:(int)groupNumber informDelegate:(BOOL)inform
+- (void)updateSelectedState:(BOOL)selected forItem:(NSString *)identifier inGroup:(NSInteger)groupNumber informDelegate:(BOOL)inform
 {
 	// This method simply updates the selected state of the item's control, maintains selectedItems, and informs the delegate.
 	// All management of dependencies (such as deselecting other selected items in a radio-selection-mode group) is performed
